requirements.txt
Django~=4.2.5
djangorestframework~=3.14.0
drf-yasg~=1.21.7
gunicorn~=21.2.0
pytest~=7.4.2
pytest-django~=4.5.2